pyproject.toml
[project]
name = "jaffle_platform"
requires-python = ">=3.11,<3.13"
version = "0.1.0"
dependencies = [
    "dagster==1.11.0",
    "dagster-dbt>=0.26.20",
    "dagster-dlt>=0.26.20",
    "dagster-duckdb>=0.26.20",
    "dagster-duckdb-pandas>=0.26.20",
    "dagster-evidence>=0.1.6",
    "dagster-postgres>=0.27.0",
    "dagster-sling>=0.26.20",
    "dbt-postgres>=1.9.0",
    "duckdb>=1.3.1",
    "jafgen>=0.4.14",
    "nltk>=3.8.1",
    "pandas>=2.2.1",
    "psycopg2-binary>=2.9.10",
    "sqlmesh>=0.198.0",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli[local]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "jaffle_platform"
registry_modules = [
    "jaffle_platform.components.*",
]

[tool.uv.sources]
dagster-sqlmesh = { path = "../dagster-sqlmesh", editable = true }


